Struct sendgrid::mail::Mail

pub struct Mail {
    pub to: Vec<String>,
    pub to_names: Vec<String>,
    pub cc: Vec<String>,
    pub bcc: Vec<String>,
    pub from: String,
    pub subject: String,
    pub html: String,
    pub text: String,
    pub from_name: String,
    pub reply_to: String,
    pub date: String,
    pub attachments: HashMap<String, String>,
    pub content: HashMap<String, String>,
    pub headers: HashMap<String, String>,
    pub x_smtpapi: String,
}

This is a representation of a valid SendGrid message. It has support for all of the fields in the V2 API.

Fields

to

to_names

cc

bcc

from

subject

html

text

from_name

reply_to

date

attachments

content

headers

x_smtpapi

Methods

impl Mail

fn new() -> Mail

Returns a new Mail struct to send with a client. All of the fields are initially empty.

fn add_cc<T: Into<String>>(&mut self, cc_addr: T)

Adds a CC recipient to the Mail struct.

fn add_to<T: Into<String>>(&mut self, to_addr: T)

Adds a to recipient to the Mail struct.

fn add_from<T: Into<String>>(&mut self, from_addr: T)

Set the from address for the Mail struct. This can be changed, but there is only one from address per message.

fn add_subject<T: Into<String>>(&mut self, subject: T)

Set the subject of the message.

fn add_html<T: Into<String>>(&mut self, html: T)

This function sets the HTML content for the message.

fn add_to_name<T: Into<String>>(&mut self, to_name: T)

Add a name for the "to" field in the message. The number of to names must match the number of "to" addresses.

fn add_text<T: Into<String>>(&mut self, text: T)

Set the text content of the message.

fn add_bcc<T: Into<String>>(&mut self, bcc_addr: T)

Add a BCC address to the message.

fn add_from_name<T: Into<String>>(&mut self, from_name: T)

Set the from name for the message.

fn add_reply_to<T: Into<String>>(&mut self, reply_to: T)

Set the reply to address for the message.

fn add_date(&mut self, date: String)

Set the date for the message. This must be a valid RFC 822 timestamp.

fn add_attachment(&mut self, path: &str)

Add an attachment for the message. You can pass the name of a file as a path on the file system.

Examples

let mut message = Mail::new();
message.add_attachment("/path/to/file/contents.txt");

fn add_content(&mut self, id: &str, value: &str)

Add content for inline images in the message.

fn add_header(&mut self, header: &str, value: &str)

Add a custom header for the message. These are usually prefixed with 'X' or 'x' per the RFC specifications.

fn make_header_string(&mut self) -> String

Used internally for string encoding. Not needed for message building.

fn add_x_smtpapi(&mut self, x_smtpapi: String)

Add an X-SMTPAPI string to the message. This can be done by using the 'rustc_serialize' crate and JSON encoding a map or custom struct. Or a regular String type can be escaped and used.