Struct sendgrid::v3::SGMailV3

pub struct SGMailV3 {
    // some fields omitted
}

The main structure for a V3 API mail send call. This is composed of many other smaller structures used to add lots of customization to your message.

Methods

impl SGMailV3

fn new() -> SGMailV3

Construct a new V3 message.

fn set_from(&mut self, from: Email)

Set the from address.

fn set_subject(&mut self, subject: &str)

Set the subject.

fn add_content(&mut self, content: Content)

Add content to the message.

fn add_personalization(&mut self, p: Personalization)

Add a personalization to the message.