Module sendgrid::v3

Structs

Attachment
An attachment block for a V3 message. Content and filename are required. If the mime_type is unspecified, the email will use Sendgrid's default for attachments which is 'application/octet-stream'.

Content
The body of an email with the content type and the message.

Email
An email with a required address and an optional name field.

Personalization
A personalization block for a V3 message. It has to at least contain one email as a to address. All other fields are optional.

Response
A Response to a submitted Request.

SGMailV3
The main structure for a V3 API mail send call. This is composed of many other smaller structures used to add lots of customization to your message.

V3Sender
Used to send a V3 message body.

Type Definitions

SGMap
Just a redefinition of a map to store string keys and values.