Struct sendgrid::v3::SGMailV3

pub struct SGMailV3 { /* fields omitted */ }

The main structure for a V3 API mail send call. This is composed of many other smaller structures used to add lots of customization to your message.

Methods

impl SGMailV3

pub fn new() -> SGMailV3

Construct a new V3 message.

pub fn set_from(&mut self, from: Email)

Set the from address.

pub fn set_subject(&mut self, subject: &str)

Set the subject.

pub fn add_content(&mut self, content: Content)

Add content to the message.

pub fn add_personalization(&mut self, p: Personalization)

Add a personalization to the message.

pub fn add_attachment(&mut self, a: Attachment)

Add an attachment to the message.